refactor: user-authentication
- authentication refactor completed.
- add permission check in admin page.
- some fixes

+++ src/main/java/com/jiezipoi/mall/config/GoodsConfig.java
@@ -69,7 +69,7 @@ public class GoodsConfig {
         return userTempFilePrefix;
     }
 
-    public String getUserTempFileName(int userName) {
+    public String getUserTempFileName(long userName) {
         return getUserTempFilePrefix() + userName;
     }
